requirements.txt
chromadb>=0.6.0
openai>=1.52.0
python-dotenv==1.0.0
aiohttp==3.9.5
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
